requirements.txt
Django
whitenoise
djangorestframework
django-cors-headers
python-dotenv 
psycopg2-binary
django-environ
djangorestframework-simplejwt
